audico_product_manager/requirements.txt

google-cloud-storage>=2.10.0
google-cloud-documentai>=2.20.1
requests>=2.28.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
pandas>=1.5.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
Pillow>=9.5.0
openai>=1.0.0
